+++ src/utils/strings.ts
@@ -1,0 +1,38 @@
+/********************************************************************************
+ * Copyright (c) 2025 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0, or the W3C Software Notice and
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR W3C-20150513
+ ********************************************************************************/
+export const extractIndexFromId = (id: string): number => {
+  const parts = id.split(" - ");
+  return parseInt(parts[1], 10);
+};
+
+export const formatTextKey = (key: string, index: number): string => {
+  let propIndex = index + 1;
+  return key + ` (opt.${String(propIndex)})`;
+};
+
+export const formatText = (text: string): string => {
+  if (text.startsWith("modbus:")) {
+    text = text.replace("modbus:", "");
+  }
+  if (text.startsWith("propName")) {
+    text = "Property Name";
+  }
+  if (text.startsWith("href")) {
+    text = "Resource";
+  }
+  if (text.startsWith("htv:methodName")) {
+    text = "Method";
+  }
+  text = text.replace(/([a-z])([A-Z])/g, "$1 $2");
+  return text.charAt(0).toUpperCase() + text.slice(1);
+};
